fix(copy+meta): audit polish — per-page titles, placeholders, button copy
From the April-16 live-site audit:

- Per-page <title> fallbacks: /login, /signup, /features, /contact,
  /api-docs, /forgot-password, /reset-password all fell back to the
  generic root title "Aries AI - Marketing Operating System". Added
  explicit metadata.title exports matching the pattern already
  shipping on /dashboard, /documentation, /privacy, /terms,
  /not-found.
- Restructured /forgot-password into the same server-component +
  page-client split the other auth routes use, so it can export
  metadata (client components can't).
- /marketing/new-job competitor-URL placeholder was still the
  audit-v2 "betterup.com" that M6 was meant to remove — that fix only
  touched the onboarding copies. Replaced with "competitor.com" to
  match the onboarding placeholder.
- /signup: "Get started with your autonomous marketing workspace" →
  "Plan, approve, and launch campaigns from one workspace." The
  review-first narrative is our differentiator; "autonomous" implies
  the opposite and conflicts with the "nothing goes live without
  your approval" promise on the homepage.
- /signup: the submit button said "REGISTER" in all caps while every
  other button in the product is sentence case — changed to "Create
  account" (and the submitting state from "Registering..." to
  "Creating account…").
