+++ app/signup/page.tsx
@@ -1,6 +1,10 @@
 import { Suspense } from 'react';
 import SignUpPageClient from './page-client';
 
+export const metadata = {
+  title: 'Create your account — Aries AI',
+};
+
 export default function SignUpPage() {
   return (
     <Suspense>
