+++ frontend/auth/sign-up-form.tsx
@@ -132,7 +132,7 @@ const SignUpForm: React.FC<SignUpFormProps> = ({
           {invitationData ? `Joining ${invitationData.organizations.name}` : "Create Account"}
         </h1>
         <p className="text-white/60">
-          Get started with your autonomous marketing workspace
+          Plan, approve, and launch campaigns from one workspace.
         </p>
       </div>
 
@@ -234,7 +234,7 @@ const SignUpForm: React.FC<SignUpFormProps> = ({
             disabled={isLoading || isSubmitting}
             className="w-full py-3 px-4 bg-gradient-to-r from-primary to-secondary hover:opacity-90 text-white font-medium rounded-xl transition-all shadow-[0_0_20px_rgba(124,58,237,0.3)] hover:shadow-[0_0_30px_rgba(124,58,237,0.5)] flex items-center justify-center gap-2 disabled:opacity-60"
           >
-            {isSubmitting ? 'Registering...' : 'REGISTER'}
+            {isSubmitting ? 'Creating account…' : 'Create account'}
           </button>
         </form>
